Make sure to also match the first aligned nucleotide
The SL sequence often (always?) ends in the same nucleotide (often
"G") as the SL acceptor site dinucleotide. This means that this "G"
is parted of the aligned portion of the sequence and is not retained
in the softclipped part. To alleviate this, we match both the
softclipped part only, and separately, the softclipped part plus the
first nucleotide of the aligned sequence. Alternatively, we could
trim a trailing "G" of the given SL sequence, but we don't know if
this would hold for all organisms.